short, long, serial change

+++ UserService/src/main/resources/liquibase/playlist/playlist-structure.sql
@@ -1,12 +1,12 @@
 CREATE TABLE playlists (
-    playlist_id SERIAL NOT NULL PRIMARY KEY,
+    playlist_id BIGSERIAL NOT NULL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     creator_id INTEGER NOT NULL
 );
 
 CREATE TABLE playlist_tracks (
     playlist_id INTEGER NOT NULL,
-    track_id LONG NOT NULL,
+    track_id BIGINT NOT NULL,
 
     PRIMARY KEY (playlist_id, track_id),
     FOREIGN KEY (playlist_id) REFERENCES playlists(playlist_id) ON DELETE CASCADE
